abolish v-menu
