+++ client/assets/variables.scss
@@ -22,7 +22,8 @@ $g-navigation-drawer-width: 200px;
  * z-index
  */
 $__layer-map: (
-  toast: 2000,
+  toast: 2001,
+  search-result-list: 2000,
   modal: 1001,
   header: 1000,
   navigation-drawer: 1000,
